% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icos_citation.R
\name{icos_citation}
\alias{icos_citation}
\title{Get a digital object's citation}
\usage{
icos_citation(dobj, style = "apa", lang = "en-GB")
}
\arguments{
\item{dobj}{a digital object URI or PID}

\item{style}{the sitation style (default = 'apa')}

\item{lang}{the language of the common citation (default = 'en-GB')}
}
\value{
a dataframe with an icos and common citation to be used in
 publications when using the referenced digital object
}
\description{
Grabs the citation of the digital object from the ICOS server or
the CrossRef server if it exists. Both this internal ICOS citation
as well as the common 'academic' citation is provided if available.
}
\examples{
\dontrun{
dobj <- "https://meta.icos-cp.eu/objects/lNJPHqvsMuTAh-3DOvJejgYc"
print(icos_citation(dobj))
}
}
