% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icos_collections.R
\name{icos_collections}
\alias{icos_collections}
\title{ICOS collections}
\usage{
icos_collections(id)
}
\arguments{
\item{id}{an URI or DOI, default (missing)}
}
\value{
all ICOS collections (or a desired subset)
}
\description{
Returns icon collections. By default all collections are returned, if an id
is given only collection information for this partuicular DOI is provided.
}
