% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icos_dobj.R
\name{icos_dobj_info}
\alias{icos_dobj_info}
\title{Grab digital object info}
\usage{
icos_dobj_info(dobj)
}
\arguments{
\item{dobj}{digital object URI or PID}
}
\value{
a data frame with digital object information
}
\description{
Get additional information on a digital object.
}
\examples{
\dontrun{
dobj <- "https://meta.icos-cp.eu/objects/lNJPHqvsMuTAh-3DOvJejgYc"
print(icos_dobj_info(dobj))
}
}
