% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icos_download.R
\name{icos_download}
\alias{icos_download}
\title{Download an ICOS digital object}
\usage{
icos_download(dobj, path = tempdir(), internal = FALSE, verbose = TRUE)
}
\arguments{
\item{dobj}{digital object identified by a persistent identifier (PID) URI}

\item{path}{where to save the downloaded data (default = tempdir())}

\item{internal}{read in data and return an internal data frame to the
current workspace}

\item{verbose}{verbose feeback on licensing, needs you to acknowledge
the license agreement for a download - setting this to FALSE overrides
this but by doing so you acknowledge the statement in bulk!}
}
\value{
a downloaded digital file on disk for further processing
}
\description{
Download a digital object identified by a persistent identifier.
}
\examples{
\dontrun{
dobj <- "https://meta.icos-cp.eu/objects/lNJPHqvsMuTAh-3DOvJejgYc"
icos_download(dobj = dobj, verbose = FALSE)
}
}
