% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icos_products.R
\name{icos_products}
\alias{icos_products}
\title{List products for a station}
\usage{
icos_products(uri, level = 2)
}
\arguments{
\item{uri}{url pointing to a resource}

\item{level}{data quality level (default = 2)}
}
\value{
a list of station meta-data
}
\description{
A list of all station data types including some
meta-data such as the estimated size, the type of data,
the time frame and the required URI for downloading
the data.
}
\examples{
\dontrun{
products <- icos_products()
}
}
