% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icos_station_data.R
\name{icos_station_data}
\alias{icos_station_data}
\title{Station data}
\usage{
icos_station_data(uri, level = 2)
}
\arguments{
\item{uri}{url pointing to a resource}

\item{level}{data quality level (default = 2)}
}
\value{
a list of station meta-data
}
\description{
A list of all station data types
}
\examples{
\dontrun{
all_station_info <- icos_stations()
}
}
