% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icos_stations.R
\name{icos_stations}
\alias{icos_stations}
\title{List all ICOS collections}
\usage{
icos_stations(station)
}
\arguments{
\item{station}{station id to subset from larger list (default missing)}
}
\value{
a data frame with ICOS collections
}
\description{
A list of all ICOS collections. Collections are aggregate products, you
will find additional identifying DOIs for proper referencing and further
exploration in the returned table.
}
\examples{
\dontrun{
station_info <- icos_stations()
print(station_info)
}
}
