% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icos_stations.R
\name{icos_stations}
\alias{icos_stations}
\title{List all ICOS station info}
\usage{
icos_stations(station)
}
\arguments{
\item{station}{station id to subset from larger list (default missing)}
}
\value{
a list of station meta-data
}
\description{
A list of all station meta-data such as latittude, longitude and elevation
}
\examples{
\dontrun{
all_station_info <- icos_stations()
}
}
